Initial Commit
Just now starting up the creation of this new project. I am attempting
to learn more about using git and github so please bare with me as I
attempt. Any help is appreciated greatly.

If anyone would like to assist on this project or any other project on
my github I am more than happy for the help as I am working on these to
help anybody that wants to work on game design.